+++ train_sweep.py
@@ -6,7 +6,7 @@ from tqdm.auto import tqdm
 import wandb
 import yaml
 from dataset import OTFDataset, Subset, collate_fn
-from modules import RNNModel, CNNModel
+from modules.baselines import RNNModel, CNNModel
 from data_utils import get_binary_f1, get_multiclass_acc
 import constants as C
